A few clean ups
